pyproject.toml
[project]
name = "trading-agent"
version = "0.1.0"
description = "AI-powered short-term options trading bot using Bill Fanter's supply/demand methodology"
requires-python = ">=3.11"
dependencies = [
    # Core
    "asyncio>=3.4.3",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "httpx>=0.26.0",

    # Data processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "ta-lib>=0.4.28",

    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",

    # Redis
    "redis>=5.0.0",
    "aioredis>=2.0.0",

    # LLM
    "anthropic>=0.39.0",

    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",

    # Monitoring & Logging
    "structlog>=24.1.0",
    "rich>=13.7.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Timezone
    "pytz>=2024.1",

    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[project.optional-dependencies]
transcriber = [
    "yt-dlp>=2024.1.0",
    "openai-whisper>=20231117",
    "openai>=1.0.0",
]
dev = [
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
